# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Travel Booking API",
        "description": "API documentation for the Travel Booking system",
        "contact": {
            "email": "hey@api-travel.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/bookings/{booking}/passengers": {
            "get": {
                "tags": [
                    "Passengers"
                ],
                "summary": "List passengers for a booking",
                "operationId": "d1fc4d7ad145365da8c4927957c17273",
                "parameters": [
                    {
                        "name": "booking",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of passengers"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Passengers"
                ],
                "summary": "Add a passenger to a booking",
                "operationId": "17a7cb6e6c940ca538113fdc4b16bed2",
                "parameters": [
                    {
                        "name": "booking",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "birth_date"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Ali"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Rezaei"
                                    },
                                    "national_id": {
                                        "type": "string",
                                        "example": "1234567890",
                                        "nullable": true
                                    },
                                    "passport_number": {
                                        "type": "string",
                                        "example": "A12345678",
                                        "nullable": true
                                    },
                                    "birth_date": {
                                        "type": "string",
                                        "example": "1990-05-15"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Passenger added"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/bookings/{booking}/passengers/{passenger}": {
            "delete": {
                "tags": [
                    "Passengers"
                ],
                "summary": "Remove a passenger from a booking",
                "operationId": "d38bdc6a3948ef7e20f437110e7d7937",
                "parameters": [
                    {
                        "name": "booking",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "passenger",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Passenger removed"
                    },
                    "422": {
                        "description": "Cannot remove passenger with issued ticket"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/bookings/{booking}/payment": {
            "get": {
                "tags": [
                    "Payments"
                ],
                "summary": "Get payment for a booking",
                "operationId": "d2e8520a3169da359693bd9fa70a567b",
                "parameters": [
                    {
                        "name": "booking",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment details"
                    },
                    "404": {
                        "description": "No payment found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Process payment for a booking",
                "operationId": "a09d2f1ffe23c1491d333e6026c5dafd",
                "parameters": [
                    {
                        "name": "booking",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "method"
                                ],
                                "properties": {
                                    "method": {
                                        "type": "string",
                                        "enum": [
                                            "card",
                                            "cash",
                                            "online"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Payment processed"
                    },
                    "422": {
                        "description": "Already paid or booking cancelled"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/bookings/{booking}/passengers/{passenger}/ticket": {
            "get": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Get ticket for a passenger",
                "operationId": "5fa8cbc782c84d3ddc3488e2f88472ed",
                "parameters": [
                    {
                        "name": "booking",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "passenger",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ticket details"
                    },
                    "404": {
                        "description": "No ticket issued yet"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/bookings/{booking}/passengers/{passenger}/ticket/issue": {
            "post": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Issue a ticket for a passenger",
                "operationId": "f63af34d859e2ce32bf0cfc07e52b395",
                "parameters": [
                    {
                        "name": "booking",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "passenger",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Ticket issued"
                    },
                    "422": {
                        "description": "Booking not confirmed or ticket already issued"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format: Bearer <token>",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}